API improvements
* Replace `Server::on_write` with `Server::on_write_ext`
* Move `GattError` from the `gatt_client` module to the `types` module
* Improve `DeferredReply::reply` to take a `Result`
