+++ nrf-softdevice/src/ble/gatt_client.rs
@@ -1,7 +1,6 @@
 //! Generic Attribute client. GATT clients consume functionality offered by GATT servers.
 
 use heapless::Vec;
-use num_enum::{FromPrimitive, IntoPrimitive};
 
 use crate::ble::*;
 use crate::util::{get_flexarray, get_union_field, Portal};
@@ -43,43 +42,6 @@ pub trait Client {
     /// If no error is returned, this instance is considered ready to use and is returned to
     /// the caller of [`discover`]
     fn discovery_complete(&mut self) -> Result<(), DiscoverError>;
-}
-
-#[rustfmt::skip]
-#[repr(u32)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-#[derive(Debug, PartialEq, Eq, Clone, Copy, IntoPrimitive, FromPrimitive)]
-pub enum GattError {
-    // This is not really an error, but IMO it's better to add it
-    // anyway, just in case someone mistakenly converts BLE_GATT_STATUS_SUCCESS into GattError.
-    // if they see "Success" they'll easily realize their mistake, if they see "Unknown" it'd be confusing.
-    Success = raw::BLE_GATT_STATUS_SUCCESS,
-
-    #[num_enum(default)]
-    Unknown = raw::BLE_GATT_STATUS_UNKNOWN,
-
-    AtterrInvalid = raw::BLE_GATT_STATUS_ATTERR_INVALID,
-    AtterrInvalidHandle = raw::BLE_GATT_STATUS_ATTERR_INVALID_HANDLE,
-    AtterrReadNotPermitted = raw::BLE_GATT_STATUS_ATTERR_READ_NOT_PERMITTED,
-    AtterrWriteNotPermitted = raw::BLE_GATT_STATUS_ATTERR_WRITE_NOT_PERMITTED,
-    AtterrInvalidPdu = raw::BLE_GATT_STATUS_ATTERR_INVALID_PDU,
-    AtterrInsufAuthentication = raw::BLE_GATT_STATUS_ATTERR_INSUF_AUTHENTICATION,
-    AtterrRequestNotSupported = raw::BLE_GATT_STATUS_ATTERR_REQUEST_NOT_SUPPORTED,
-    AtterrInvalidOffset = raw::BLE_GATT_STATUS_ATTERR_INVALID_OFFSET,
-    AtterrInsufAuthorization = raw::BLE_GATT_STATUS_ATTERR_INSUF_AUTHORIZATION,
-    AtterrPrepareQueueFull = raw::BLE_GATT_STATUS_ATTERR_PREPARE_QUEUE_FULL,
-    AtterrAttributeNotFound = raw::BLE_GATT_STATUS_ATTERR_ATTRIBUTE_NOT_FOUND,
-    AtterrAttributeNotLong = raw::BLE_GATT_STATUS_ATTERR_ATTRIBUTE_NOT_LONG,
-    AtterrInsufEncKeySize = raw::BLE_GATT_STATUS_ATTERR_INSUF_ENC_KEY_SIZE,
-    AtterrInvalidAttValLength = raw::BLE_GATT_STATUS_ATTERR_INVALID_ATT_VAL_LENGTH,
-    AtterrUnlikelyError = raw::BLE_GATT_STATUS_ATTERR_UNLIKELY_ERROR,
-    AtterrInsufEncryption = raw::BLE_GATT_STATUS_ATTERR_INSUF_ENCRYPTION,
-    AtterrUnsupportedGroupType = raw::BLE_GATT_STATUS_ATTERR_UNSUPPORTED_GROUP_TYPE,
-    AtterrInsufResources = raw::BLE_GATT_STATUS_ATTERR_INSUF_RESOURCES,
-    AtterrCpsWriteReqRejected = raw::BLE_GATT_STATUS_ATTERR_CPS_WRITE_REQ_REJECTED,
-    AtterrCpsCccdConfigError = raw::BLE_GATT_STATUS_ATTERR_CPS_CCCD_CONFIG_ERROR,
-    AtterrCpsProcAlrInProg = raw::BLE_GATT_STATUS_ATTERR_CPS_PROC_ALR_IN_PROG,
-    AtterrCpsOutOfRange = raw::BLE_GATT_STATUS_ATTERR_CPS_OUT_OF_RANGE,
 }
 
 /// Error type for [`discover`]
