+++ nrf-softdevice/src/ble/replies.rs
@@ -157,7 +157,7 @@ struct DeferredReply<const DEFERRED_TYPE: u8> {
 impl<const DEFERRED_TYPE: u8> Drop for DeferredReply<DEFERRED_TYPE> {
     fn drop(&mut self) {
         warn!("DeferredReply<{}> dropped without reply", DEFERRED_TYPE);
-        let res = unsafe { self.finalize(raw::BLE_GATT_STATUS_ATTERR_UNLIKELY_ERROR as u16, &[], false) };
+        let res = unsafe { self.finalize(DEFERRED_TYPE, Err(super::GattError::AtterrAttributeNotFound)) };
 
         if let Err(_err) = res {
             warn!("sd_ble_gatts_rw_authorize_reply err {:?}", _err);
@@ -167,8 +167,9 @@ impl<const DEFERRED_TYPE: u8> Drop for DeferredReply<DEFERRED_TYPE> {
 
 #[cfg(feature = "ble-gatt-server")]
 impl<const DEFERRED_TYPE: u8> DeferredReply<DEFERRED_TYPE> {
-    fn reply(mut self, status: u16, data: &[u8], update: bool) -> Result<(), RawError> {
-        let res = unsafe { self.finalize(status, data.as_ref(), update) };
+    fn reply(mut self, res: Result<Option<&[u8]>, super::GattError>) -> Result<(), RawError> {
+        assert!(res != Err(super::GattError::Success));
+        let res = unsafe { self.finalize(DEFERRED_TYPE, res) };
         core::mem::forget(self);
         res
     }
@@ -176,18 +177,28 @@ impl<const DEFERRED_TYPE: u8> DeferredReply<DEFERRED_TYPE> {
     /// # Safety
     ///
     /// This method must be called exactly once
-    unsafe fn finalize(&mut self, status: u16, data: &[u8], update: bool) -> Result<(), RawError> {
+    unsafe fn finalize(
+        &mut self,
+        deferred_type: u8,
+        res: Result<Option<&[u8]>, super::GattError>,
+    ) -> Result<(), RawError> {
+        let (gatt_status, update, p_data, len) = match res {
+            Ok(Some(data)) => (super::GattError::Success, true, data.as_ptr(), data.len()),
+            Ok(None) => (super::GattError::Success, false, core::ptr::null(), 0),
+            Err(status) => (status, false, core::ptr::null(), 0),
+        };
+
         let res = if let Some(handle) = self.conn.handle() {
             let params = raw::ble_gatts_authorize_params_t {
-                gatt_status: status,
+                gatt_status: u32::from(gatt_status) as u16,
                 _bitfield_1: raw::ble_gatts_authorize_params_t::new_bitfield_1(u8::from(update)),
                 offset: 0,
-                len: data.len() as u16,
-                p_data: data.as_ptr(),
+                len: len as u16,
+                p_data,
             };
 
             let reply_params = raw::ble_gatts_rw_authorize_reply_params_t {
-                type_: DEFERRED_TYPE,
+                type_: deferred_type,
                 params: raw::ble_gatts_rw_authorize_reply_params_t__bindgen_ty_1 { read: params },
             };
 
@@ -207,6 +218,7 @@ impl<const DEFERRED_TYPE: u8> DeferredReply<DEFERRED_TYPE> {
 pub struct DeferredWriteReply(DeferredReply<DEFERRED_TYPE_WRITE>);
 
 #[cfg(feature = "ble-gatt-server")]
+/// Represents an in-progress deferred write request
 impl DeferredWriteReply {
     pub(crate) fn new(conn: Connection) -> Self {
         DeferredWriteReply(DeferredReply {
@@ -218,9 +230,8 @@ impl DeferredWriteReply {
         &self.0.conn
     }
 
-    pub fn reply<T: AsRef<[u8]>>(self, status: u16, data: &T) -> Result<(), RawError> {
-        let update = u32::from(status) == raw::BLE_GATT_STATUS_SUCCESS;
-        self.0.reply(status, data.as_ref(), update)
+    pub fn reply(self, res: Result<&[u8], super::GattError>) -> Result<(), RawError> {
+        self.0.reply(res.map(Some))
     }
 }
 
@@ -228,6 +239,7 @@ impl DeferredWriteReply {
 pub struct DeferredReadReply(DeferredReply<DEFERRED_TYPE_READ>);
 
 #[cfg(feature = "ble-gatt-server")]
+/// Represents an in-progress deferred read request
 impl DeferredReadReply {
     pub(crate) fn new(conn: Connection) -> Self {
         DeferredReadReply(DeferredReply {
@@ -239,7 +251,10 @@ impl DeferredReadReply {
         &self.0.conn
     }
 
-    pub fn reply<T: AsRef<[u8]>>(self, status: u16, data: &T, update: bool) -> Result<(), RawError> {
-        self.0.reply(status, data.as_ref(), update)
+    /// Finishes the read operation with `res`.
+    ///
+    /// If `res` is `Ok(None)`, the value of the attribute stored by the softdevice will be returned to the central.
+    pub fn reply(self, res: Result<Option<&[u8]>, super::GattError>) -> Result<(), RawError> {
+        self.0.reply(res)
     }
 }
